requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
Werkzeug==3.0.1
python-dotenv==1.0.1
gunicorn==21.2.0
